/* PR_START p/semantic_graph 14 */ Move `bookings_query_parser` fixture to `conftest.py`.
