+++ metricflow-semantics/tests_metricflow_semantics/query/test_query_parser.py
@@ -2,18 +2,14 @@ from __future__ import annotations
 
 import logging
 import textwrap
-from typing import List
 
 import pytest
 from _pytest.fixtures import FixtureRequest
 from dbt_semantic_interfaces.implementations.semantic_manifest import PydanticSemanticManifest
-from dbt_semantic_interfaces.parsing.dir_to_model import parse_yaml_files_to_validation_ready_semantic_manifest
 from dbt_semantic_interfaces.parsing.objects import YamlConfigFile
-from dbt_semantic_interfaces.protocols import SemanticManifest
 from dbt_semantic_interfaces.references import EntityReference
 from dbt_semantic_interfaces.test_utils import as_datetime
 from dbt_semantic_interfaces.type_enums.date_part import DatePart
-from dbt_semantic_interfaces.validations.semantic_manifest_validator import SemanticManifestValidator
 from metricflow_semantics.model.semantic_manifest_lookup import SemanticManifestLookup
 from metricflow_semantics.query.query_exceptions import InvalidQueryException
 from metricflow_semantics.query.query_parser import MetricFlowQueryParser
@@ -31,44 +27,9 @@ from metricflow_semantics.test_helpers.example_project_configuration import (
 from metricflow_semantics.test_helpers.metric_time_dimension import MTD
 from metricflow_semantics.test_helpers.snapshot_helpers import assert_object_snapshot_equal
 
+from tests_metricflow_semantics.query.conftest import BOOKINGS_YAML, query_parser_from_yaml
+
 logger = logging.getLogger(__name__)
-
-
-BOOKINGS_YAML = textwrap.dedent(
-    """\
-    semantic_model:
-      name: bookings_source
-
-      node_relation:
-        schema_name: some_schema
-        alias: bookings_source_table
-
-      defaults:
-        agg_time_dimension: ds
-
-      measures:
-        - name: bookings
-          expr: "1"
-          agg: sum
-          create_metric: true
-
-      dimensions:
-        - name: is_instant
-          type: categorical
-        - name: ds
-          type: time
-          type_params:
-            time_granularity: day
-
-      primary_entity: booking
-
-      entities:
-        - name: listing
-          type: foreign
-          expr: listing_id
-    """
-)
-
 
 REVENUE_YAML = textwrap.dedent(
     """\
@@ -177,12 +138,6 @@ METRICS_YAML = textwrap.dedent(
             alias: revenue_last_7_days
     """
 )
-
-
-@pytest.fixture
-def bookings_query_parser() -> MetricFlowQueryParser:  # noqa
-    bookings_yaml_file = YamlConfigFile(filepath="inline_for_test_1", contents=BOOKINGS_YAML)
-    return query_parser_from_yaml([EXAMPLE_PROJECT_CONFIGURATION_YAML_CONFIG_FILE, bookings_yaml_file])
 
 
 @pytest.fixture
@@ -600,18 +555,6 @@ def test_offset_metric_with_diff_agg_time_dims_error() -> None:  # noqa: D103
             metric_names=["monthly_revenue_last_7_days"],
             group_by_names=["revenue___ds"],
         )
-
-
-def query_parser_from_yaml(yaml_contents: List[YamlConfigFile]) -> MetricFlowQueryParser:
-    """Given yaml files, return a query parser using default source nodes, resolvers and time spine source."""
-    semantic_manifest = parse_yaml_files_to_validation_ready_semantic_manifest(
-        yaml_contents, apply_transformations=True
-    ).semantic_manifest
-    semantic_manifest_lookup = SemanticManifestLookup(semantic_manifest)
-    SemanticManifestValidator[SemanticManifest]().checked_validations(semantic_manifest)
-    return MetricFlowQueryParser(
-        semantic_manifest_lookup=semantic_manifest_lookup,
-    )
 
 
 def test_invalid_group_by_metric(bookings_query_parser: MetricFlowQueryParser) -> None:
